Split IncidentController into a few files.
